more on types and casts
git-svn-id: https://svn.r-project.org/R/trunk@59097 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/rcont.c
@@ -70,7 +70,7 @@ void rcont2(int *nrow, int *ncol,
 
                 /* Compute conditional expected value of MATRIX(L, M) */
 
-                nlm = ia * (id / (double)ie) + 0.5;
+                nlm = (int)(ia * (id / (double)ie) + 0.5);
                 x = exp(fact[ia] + fact[ib] + fact[ic] + fact[id] - fact[ie] - fact[nlm] - fact[id - nlm] -
                         fact[ia - nlm] - fact[ii + nlm]);
                 if (x >= dummy)
@@ -85,7 +85,7 @@ void rcont2(int *nrow, int *ncol,
                 do
                 {
                     /* Increment entry in row L, column M */
-                    j = (id - nlm) * (double)(ia - nlm);
+                    j = (int)((id - nlm) * (double)(ia - nlm));
                     lsp = (j == 0);
                     if (!lsp)
                     {
@@ -101,7 +101,7 @@ void rcont2(int *nrow, int *ncol,
                         R_CheckUserInterrupt();
 
                         /* Decrement entry in row L, column M */
-                        j = nll * (double)(ii + nll);
+                        j = (int)(nll * (double)(ii + nll));
                         lsm = (j == 0);
                         if (!lsm)
                         {
